pyproject.toml
[project]
name = "event-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backports-zoneinfo==0.2.1",
    "django==4.2.23",
    "djangorestframework==3.15.2",
    "drf-spectacular==0.28.0",
    "importlib-resources==6.4.5",
    "inflection==0.5.1",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "pkgutil-resolve-name==1.3.10",
    "psycopg2-binary==2.9.10",
    "pyyaml==6.0.2",
    "referencing==0.35.1",
    "rpds-py==0.20.1",
    "sqlparse==0.5.3",
    "typing-extensions==4.13.2",
    "uritemplate==4.1.1",
    "zipp==3.20.2",
]
